first round of cleanup
